[#12427] test(iceberg): move the schema lifecycle into the base class
All three subclasses had the same create-if-absent and teardown blocks.
The base class already owns SCHEMA_NAME and clearTable(), so it owns the
drop as well.
